requirements.txt
alembic==0.9.9
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
httplib2==0.11.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
Mako==1.0.7
MarkupSafe==1.1.0
meteostat==0.2.0
numpy==1.19.4
oauth2==1.9.0.post1
oauthlib==2.0.7
pandas==1.1.4
psycopg2-binary==2.8.6
pyarrow==2.0.0
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2020.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
urllib3==1.22
URLObject==2.4.3
Werkzeug==0.14.1
wincertstore==0.2
WTForms==2.1
